+++ src/cqterrain/segment/__init__.py
@@ -1,0 +1,1 @@
+from .BaseSegment import BaseSegment